Add files via upload
A diagram of the flow of the frontend we would like to have. We want to have the home page then have 5 different tabs. Four of the tabs will be different categories of machines so users can find the machines they are looking for faster. Then the last tab will be setting tab for the admins to use so they can change settings of the website.